refactor(motion): split actions, activities, and constraints into separate modules
