+++ unified_planning/model/motion/__init__.py
@@ -13,15 +13,14 @@
 # limitations under the License.
 #
 
-from unified_planning.model.tamp.action import (
+from unified_planning.model.motion.action import (
     InstantaneousMotionAction,
     DurativeMotionAction,
-    MotionActivity,
     MotionConstraint,
-    Waypoints,
-    ActivityWaypoints,
 )
-from unified_planning.model.tamp.objects import (
+from unified_planning.model.motion.activity import MotionActivity
+from unified_planning.model.motion.constraint import Waypoints, ActivityWaypoints
+from unified_planning.model.motion.objects import (
     ConfigurationKind,
     ConfigurationInstance,
     SE2,
@@ -32,9 +31,14 @@ from unified_planning.model.tamp.objects import (
     MovableObject,
     ConfigurationObject,
 )
-from unified_planning.model.tamp.types import OccupancyMap
-from unified_planning.model.tamp.path import Path, SE2Path, SE3Path, SE2WithControlsPath
-from unified_planning.model.tamp.scheduling_motion_problem import (
+from unified_planning.model.motion.types import OccupancyMap
+from unified_planning.model.motion.path import (
+    Path,
+    SE2Path,
+    SE3Path,
+    SE2WithControlsPath,
+)
+from unified_planning.model.motion.scheduling_motion_problem import (
     SchedulingMotionProblem,
 )
