+++ unified_planning/model/motion/action.py
@@ -15,344 +15,14 @@
 
 
 import unified_planning as up
-from unified_planning.environment import Environment, get_environment
+from unified_planning.environment import Environment
 from unified_planning.model import InstantaneousAction, DurativeAction
 from unified_planning.model.mixins.timed_conds_effs import TimedCondsEffs
-from unified_planning.exceptions import UPTypeError, UPUnboundedVariablesError
-from abc import ABC, abstractmethod
-from typing import Optional, List, Iterable, Dict, Union, Tuple
-from collections import Counter, OrderedDict
+from typing import Optional, List, Iterable, Dict
+from collections import OrderedDict
 
-from unified_planning.model.timing import EndTiming, StartTiming, TimeInterval, Timing
-from unified_planning.model.scheduling.activity import Activity
-
-
-class Attachment(ABC):
-    """
-    This class represents an attachment between two objects.
-
-    This is defined by the involved movable objects, their attachment configurations,
-    and the associated links that establish their connection.
-
-    Once the two objects are attached, a set of touchable links is defined for each object.
-    These links are the ones for which collision checking is disabled between the two objects.
-
-    """
-
-    def __init__(
-        self,
-        activation_condition: List["up.model.fnode.FNode"],
-        obj_from: "up.model.expression.Expression",
-        obj_to: "up.model.expression.Expression",
-        attached_config_from: "up.model.expression.Expression",
-        attached_config_to: "up.model.expression.Expression",
-        attached_link_from: str,
-        attached_link_to: str,
-        touchable_links_from: Optional[List[str]] = None,
-        touchable_links_to: Optional[List[str]] = None,
-        environment: Optional[Environment] = None,
-    ):
-        self._environment = get_environment(environment)
-
-        (activation_cond_exp,) = self._environment.expression_manager.auto_promote(
-            activation_condition
-        )
-        assert self._environment.type_checker.get_type(
-            activation_cond_exp
-        ).is_bool_type()
-
-        free_vars = self._environment.free_vars_oracle.get_free_variables(
-            activation_cond_exp
-        )
-        if len(free_vars) != 0:
-            raise UPUnboundedVariablesError(
-                f"The activation condition {str(activation_cond_exp)} has unbounded variables:\n{str(free_vars)}"
-            )
-
-        (
-            obj_from_exp,
-            obj_to_exp,
-        ) = self._environment.expression_manager.auto_promote(obj_from, obj_to)
-        tm1 = self._environment.type_checker.get_type(obj_from_exp)
-        tm2 = self._environment.type_checker.get_type(obj_to_exp)
-
-        if not tm1.is_movable_type() or not tm2.is_movable_type():
-            raise UPTypeError(
-                "Objects of Transform's constructor must be of movable type!"
-            )
-
-        (
-            attached_config_from_exp,
-            attached_config_to_exp,
-        ) = self._environment.expression_manager.auto_promote(
-            attached_config_from, attached_config_to
-        )
-        tc1 = self._environment.type_checker.get_type(attached_config_from_exp)
-        tc2 = self._environment.type_checker.get_type(attached_config_to_exp)
-        if not tc1.is_configuration_type() or not tc2.is_configuration_type():
-            raise UPTypeError(
-                "Configurations of Transform's constructor must be of configuration type!"
-            )
-
-        if tc1.kind != tc2.kind:
-            raise UPTypeError(
-                "Configurations of Object 1 and 2 must be of the same configuration kind!"
-            )
-
-        self._obj_from = obj_from_exp
-        self._obj_to = obj_to_exp
-        self._attached_config_from = attached_config_from_exp
-        self._attached_config_to = attached_config_to_exp
-        self._attached_link_from = attached_link_from
-        self._attached_link_to = attached_link_to
-        self._touchable_links_from = touchable_links_from
-        self._touchable_links_to = touchable_links_to
-
-    def __eq__(self, oth) -> bool:
-        if not isinstance(oth, Attachment) or self._environment != oth._environment:
-            return False
-        if (
-            self._obj_from != oth._obj_from
-            or self._obj_to != oth._obj_to
-            or self._attached_config_from != oth._attached_config_from
-            or self._attached_config_to != oth._attached_config_to
-            or self._attached_link_from != oth._attached_link_from
-            or self._attached_link_to != oth._attached_link_to
-            or Counter(self._touchable_links_from)
-            != Counter(self._touchable_links_from)
-            or Counter(self._touchable_links_to) != Counter(self._touchable_links_to)
-        ):
-            return False
-        return True
-
-    def __hash__(self) -> int:
-        res = hash(self._obj_from)
-        res += hash(self._obj_to)
-        res += hash(self._attached_config_from)
-        res += hash(self._attached_config_to)
-        res += hash(self._attached_link_from)
-        res += hash(self._attached_link_to)
-        for l in self._touchable_links_from:
-            res += hash(l)
-        for l in self._touchable_links_to:
-            res += hash(l)
-        return res
-
-    def __repr__(self) -> str:
-        s = ["attachment("]
-        s.append(str(self.obj_from))
-        s.append(", ")
-        s.append(str(self.obj_to))
-        s.append(", ")
-        s.append(str(self.attached_config_from))
-        s.append(", ")
-        s.append(str(self.attached_config_to))
-        s.append(", ")
-        s.append(self.attached_link_from)
-        s.append(", ")
-        s.append(self.attached_link_to)
-        s.append(", ")
-        s.append("[" + ", ".join(self.touchable_links_from) + "]")
-        s.append(", ")
-        s.append("[" + ", ".join(self.touchable_links_to) + "]")
-        s.append(")")
-        return "".join(s)
-
-    @property
-    def obj_from(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the movable object to be attached."""
-        return self._obj_from
-
-    @property
-    def obj_to(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the movable object to which attach."""
-        return self._obj_to
-
-    @property
-    def attached_config_from(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the attached configuration of movable object to be attached."""
-        return self._attached_config_from
-
-    @property
-    def attached_config_to(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the attached configuration of the movable object to which attach."""
-        return self._attached_config_to
-
-    @property
-    def attached_link_from(self) -> str:
-        """Returns the name of the attached link of movable object to be attached."""
-        return self._attached_link_from
-
-    @property
-    def attached_link_to(self) -> str:
-        """Returns the name of the attached link of the movable object to which attach."""
-        return self._attached_link_to
-
-    @property
-    def touchable_links_from(self) -> List[str]:
-        """Returns the name of the touchable links of the movable object to be attached."""
-        return self._touchable_links_from
-
-    @property
-    def touchable_links_to(self) -> List[str]:
-        """Returns the name of the touchable links of the movable object to which attach."""
-        return self._touchable_links_to
-
-
-class MotionConstraint(ABC):
-    """
-    This class represents a motion constraint.
-
-    A motion constraint is a constraint that must hold true among the continuous parameters of a motion action
-    for it to be a legal transition of the system in its workspace.
-    """
-
-    def __init__(self, environment: Optional[Environment] = None):
-        self._environment = get_environment(environment)
-
-    @abstractmethod
-    def __eq__(self, oth) -> bool:
-        raise NotImplementedError
-
-    @abstractmethod
-    def __hash__(self) -> int:
-        raise NotImplementedError
-
-    @property
-    def environment(self) -> Environment:
-        """Returns the `Environment` in which this MotionConstraint exists."""
-        return self._environment
-
-
-class Waypoints(MotionConstraint):
-    """
-    This class represents a waypoints contraint.
-
-    The waypoints constraint is a `MotionConstraint` representing the existence of a trajectory
-    in the free configuration space of a movable object that lets it traverse a set of input waypoints
-    starting from an initial configuration.
-
-    `static_obstacles` is a dictionary that maps those movable obstacles that remain static during the existence of the constraint with their configurations.
-    `dynamic_obstacles_at_start` is a dictionary that maps those movable obstacles that may move during the existence of the constraint with their configurations at the beginning of the constraint.
-    `attachments` is a list of `Attachments`, i.e., extensions of movable objects that change their geometric models.
-    """
-
-    def __init__(
-        self,
-        movable: "up.model.expression.Expression",
-        starting: "up.model.expression.Expression",
-        waypoints: List["up.model.expression.Expression"],
-        static_obstacles: Optional[
-            Dict["up.model.motion.MovableObject", "up.model.fnode.FNode"]
-        ] = None,
-        dynamic_obstacles_at_start: Optional[
-            Dict["up.model.motion.MovableObject", "up.model.fnode.FNode"]
-        ] = None,
-        attachments: Optional[List[Attachment]] = None,
-        environment: Optional[Environment] = None,
-    ):
-        super().__init__(environment)
-        (movable_exp,) = self._environment.expression_manager.auto_promote(movable)
-        if not self._environment.type_checker.get_type(movable_exp).is_movable_type():
-            raise UPTypeError(
-                "First parameter of Waypoints's constructor must be of movable type!"
-            )
-        (
-            starting_exp,
-            *waypoints_exp,
-        ) = self._environment.expression_manager.auto_promote(starting, *waypoints)
-        t = self._environment.type_checker.get_type(starting_exp)
-        if not t.is_configuration_type():
-            raise UPTypeError(
-                "starting parameter of Waypoints's constructor must be of configuration type!"
-            )
-        for p in waypoints_exp:
-            pt = self._environment.type_checker.get_type(p)
-            if not pt.is_configuration_type():
-                raise UPTypeError(
-                    "waypoints parameter of Waypoints's constructor must be a list of configuration type objects!"
-                )
-            if t != pt:
-                raise UPTypeError(
-                    "starting and waypoints must be of the same configuration type!"
-                )
-        self._movable = movable_exp
-        self._starting = starting_exp
-        self._waypoints = waypoints_exp
-        self._static_obstacles = static_obstacles
-        self._dynamic_obstacles_at_start = dynamic_obstacles_at_start
-        self._attachments = attachments
-
-    def __eq__(self, oth) -> bool:
-        if not isinstance(oth, Waypoints) or self._environment != oth._environment:
-            return False
-        if self._movable != oth._movable or self._starting != oth._starting:
-            return False
-        if set(self._waypoints) != set(oth._waypoints):
-            return False
-        if self._attachments != oth._attachments:
-            return False
-        return True
-
-    def __hash__(self) -> int:
-        res = hash(self._movable)
-        res += hash(self._starting)
-        for p in self._waypoints:
-            res += hash(p)
-        return res
-
-    def __repr__(self) -> str:
-        s = ["waypoints("]
-        s.append(str(self.movable))
-        s.append(", ")
-        s.append(str(self.starting))
-        s.append(", ")
-        s.append(str(self.waypoints))
-        if self.static_obstacles is not None:
-            s.append(", ")
-            s.append(str(self.static_obstacles))
-        if self.dynamic_obstacles_at_start is not None:
-            s.append(", ")
-            s.append(str(self.dynamic_obstacles_at_start))
-        if self.attachments is not None:
-            s.append(", ")
-            s.append(str(self.attachments))
-        s.append(")")
-        return "".join(s)
-
-    @property
-    def movable(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the involved movable object."""
-        return self._movable
-
-    @property
-    def starting(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the starting configuration of the involved movable object."""
-        return self._starting
-
-    @property
-    def waypoints(self) -> List["up.model.fnode.FNode"]:
-        """Returns the `list` of `FNode` representing the set of waypoints that the involved movable object should traverse."""
-        return self._waypoints
-
-    @property
-    def static_obstacles(
-        self,
-    ) -> Optional[Dict["up.model.motion.MovableObject", "up.model.fnode.FNode"]]:
-        """Returns the set of `MovableObject` associated with the fluent expressions that represent their configuration during all the constraint (static obstacles)."""
-        return self._static_obstacles
-
-    @property
-    def dynamic_obstacles_at_start(
-        self,
-    ) -> Optional[Dict["up.model.motion.MovableObject", "up.model.fnode.FNode"]]:
-        """Returns the set of `MovableObject` associated with the fluent expressions that represent their configuration at the beginning of the constraint (possibly dynamic obstacles)."""
-        return self._dynamic_obstacles_at_start
-
-    @property
-    def attachments(self) -> List[Attachment]:
-        """Returns the list `Attachment` for this motion constraint."""
-        return self._attachments
+from unified_planning.model.motion.constraint import MotionConstraint
+from unified_planning.model.timing import EndTiming, StartTiming, TimeInterval
 
 
 class InstantaneousMotionAction(InstantaneousAction):
@@ -529,224 +199,3 @@ class DurativeMotionAction(DurativeAction):
             self._timed_motion_constraints[interval].append(motion_constraint)
         else:
             self._timed_motion_constraints[interval] = [motion_constraint]
-
-
-class ActivityWaypoints(MotionConstraint):
-
-    def __init__(
-        self,
-        movable: "up.model.expression.Expression",
-        starting: "up.model.expression.Expression",
-        waypoints: List["up.model.expression.Expression"],
-        static_obstacles: Optional[
-            Union[
-                List["up.model.motion.MovableObject"],
-                Dict["up.model.motion.MovableObject", "up.model.Fluent"],
-            ]
-        ] = None,
-        dynamic_obstacles_at_start: Optional[
-            Union[
-                List["up.model.motion.MovableObject"],
-                Dict["up.model.motion.MovableObject", "up.model.Fluent"],
-            ]
-        ] = None,
-        attachments: Optional[List[Attachment]] = None,
-        environment: Optional[Environment] = None,
-    ):
-        super().__init__(environment)
-        (movable_exp,) = self._environment.expression_manager.auto_promote(movable)
-        if not self._environment.type_checker.get_type(movable_exp).is_movable_type():
-            raise UPTypeError(
-                "First parameter of Waypoints's constructor must be of movable type!"
-            )
-        (
-            starting_exp,
-            *waypoints_exp,
-        ) = self._environment.expression_manager.auto_promote(starting, *waypoints)
-        t = self._environment.type_checker.get_type(starting_exp)
-        if not t.is_configuration_type():
-            raise UPTypeError(
-                "starting parameter of Waypoints's constructor must be of configuration type!"
-            )
-        for p in waypoints_exp:
-            pt = self._environment.type_checker.get_type(p)
-            if not pt.is_configuration_type():
-                raise UPTypeError(
-                    "waypoints parameter of Waypoints's constructor must be a list of configuration type objects!"
-                )
-            if t != pt:
-                raise UPTypeError(
-                    "starting and waypoints must be of the same configuration type!"
-                )
-        self._movable = movable_exp
-        self._starting = starting_exp
-        self._waypoints = waypoints_exp
-        self._static_obstacles = static_obstacles
-        self._dynamic_obstacles_at_start = dynamic_obstacles_at_start
-        self._attachments = attachments
-
-    def __eq__(self, oth) -> bool:
-        if not isinstance(oth, Waypoints) or self._environment != oth._environment:
-            return False
-        if self._movable != oth._movable or self._starting != oth._starting:
-            return False
-        if set(self._waypoints) != set(oth._waypoints):
-            return False
-        if self._attachments != oth._attachments:
-            return False
-        return True
-
-    def __hash__(self) -> int:
-        res = hash(self._movable)
-        res += hash(self._starting)
-        for p in self._waypoints:
-            res += hash(p)
-        return res
-
-    def __repr__(self) -> str:
-        s = ["waypoints("]
-        s.append(str(self.movable))
-        s.append(", ")
-        s.append(str(self.starting))
-        s.append(", ")
-        s.append(str(self.waypoints))
-        if self.static_obstacles is not None:
-            s.append(", ")
-            s.append(str(self.static_obstacles))
-        if self.dynamic_obstacles_at_start is not None:
-            s.append(", ")
-            s.append(str(self.dynamic_obstacles_at_start))
-        if self.attachments is not None:
-            s.append(", ")
-            s.append(str(self.attachments))
-        s.append(")")
-        return "".join(s)
-
-    @property
-    def movable(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the involved movable object."""
-        return self._movable
-
-    @property
-    def starting(self) -> "up.model.fnode.FNode":
-        """Returns the `FNode` representing the starting configuration of the involved movable object."""
-        return self._starting
-
-    @property
-    def waypoints(self) -> List["up.model.fnode.FNode"]:
-        """Returns the `list` of `FNode` representing the set of waypoints that the involved movable object should traverse."""
-        return self._waypoints
-
-    @property
-    def static_obstacles(
-        self,
-    ) -> Optional[
-        Union[
-            List["up.model.motion.MovableObject"],
-            Dict["up.model.motion.MovableObject", "up.model.Fluent"],
-        ]
-    ]:
-        return self._static_obstacles
-
-    @property
-    def dynamic_obstacles_at_start(
-        self,
-    ) -> Optional[
-        Union[
-            List["up.model.motion.MovableObject"],
-            Dict["up.model.motion.MovableObject", "up.model.Fluent"],
-        ]
-    ]:
-        return self._dynamic_obstacles_at_start
-
-    @property
-    def attachments(self) -> List[Attachment]:
-        """Returns the list `Attachment` for this motion constraint."""
-        return self._attachments
-
-
-# FIXME: align with optional-activities
-class MotionActivity(Activity):
-    """This class represents a motion activity."""
-
-    def __init__(
-        self,
-        name: str,
-        duration: int = 1,
-        optional: bool = False,
-        _env: Optional[Environment] = None,
-    ):
-        Activity.__init__(self, name, duration, optional, _env)
-        self._motion_constraints: List[MotionConstraint] = []
-        self._motion_effects: Dict["up.model.FNode", "up.model.FNode"] = {}
-
-    def __eq__(self, oth: object) -> bool:
-        if not isinstance(oth, MotionActivity):
-            return False
-        if self._motion_constraints != oth._motion_constraints:
-            return False
-        if self._motion_effects != oth._motion_effects:
-            return False
-        return super().__eq__(oth)
-
-    def __hash__(self) -> int:
-        res = super().__hash__()
-        res += sum(map(hash, self.motion_constraints))
-        res += sum(map(hash, self.motion_effects))
-        return res
-
-    def __repr__(self) -> str:
-        s = ["motion-activity {"]
-        s.append(super().__repr__().replace("/n", "\n  "))
-        s.append("  motion-constraints")
-        s.append("  " + repr(self.motion_constraints))
-        s.append("  motion-effects")
-        s.append("  " + repr(self.motion_effects))
-        s.append("}")
-        return "\n".join(s)
-
-    def clone(self):
-        new = MotionActivity(self.name, optional=self.optional, _env=self._environment)
-        self._clone_to(new)
-        new._duration = self._duration
-        new._motion_constraints = self.motion_constraints
-        new._motion_effects = self.motion_effects
-        return new
-
-    @property
-    def motion_constraints(self) -> List[MotionConstraint]:
-        return self._motion_constraints
-
-    @property
-    def motion_effects(
-        self,
-    ) -> Dict["up.model.FNode", "up.model.FNode"]:
-        return self._motion_effects
-
-    def clear_motion_constraints(self):
-        """Removes all `motion_constraints`."""
-        self._motion_constraints = []
-
-    def clear_motion_constraints(self):
-        """Removes all `motion_effects`."""
-        self._motion_effects = {}
-
-    def add_motion_constraint(self, motion_constraint: MotionConstraint):
-        self._motion_constraints.append(motion_constraint)
-
-    def add_motion_effect(
-        self,
-        movable: "up.model.expression.Expression",
-        config: "up.model.expression.Expression",
-    ):
-        (movable_exp, config_exp) = self._environment.expression_manager.auto_promote(
-            movable, config
-        )
-        if not self._environment.type_checker.get_type(movable_exp).is_movable_type():
-            raise UPTypeError("movable parameter must be of movable type!")
-        if not self._environment.type_checker.get_type(
-            config_exp
-        ).is_configuration_type():
-            raise UPTypeError("config parameter must be of configuration type!")
-
-        self._motion_effects[movable_exp] = config_exp
